code synchronization at 2017-12-01@20:45
New Features:
1. support new platform: ESP32
2. AT framework, including: AT Parser, AT HAL
3. Socket Adaption Layer(IP Mode): providing socket APIs over AT
4. 3 BINS: kernel/framwork/app
5. JavaScript Engine
6. Testbed infrastruce, working with AlisOS Studio IDE to provide
   enhanced multi-device develop/debug capability

+++ include/hal/soc/gpio.h
@@ -97,7 +97,7 @@ int32_t hal_gpio_output_toggle(gpio_dev_t *gpio);
  *
  * @return  0 : on success, EIO : if an error occurred with any step
  */
-int8_t hal_gpio_input_get(gpio_dev_t *gpio, uint32_t *value);
+int32_t hal_gpio_input_get(gpio_dev_t *gpio, uint32_t *value);
 
 /**
  * Enables an interrupt trigger for an input GPIO pin.
